Fix MSVC warning C4267
  C4267: 'argument': conversion from 'size_t' to 'const Key', possible loss of data

in

   deferredStates()->insert(qHash(object, property), state);

and

   QQmlComponentPrivate::DeferredState *state = deferredStates()->take(qHash(object, property));

The return type of qHash got changed to size_t in commit bd5b9cae7a29

+++ src/quicktemplates2/qquickdeferredexecute.cpp
@@ -48,7 +48,7 @@ QT_BEGIN_NAMESPACE
 
 namespace QtQuickPrivate {
 
-typedef QHash<uint, QQmlComponentPrivate::DeferredState *> DeferredStates;
+typedef QHash<size_t, QQmlComponentPrivate::DeferredState *> DeferredStates;
 
 static inline size_t qHash(QObject *object, const QString &propertyName)
 {
